requirements.txt
paho-mqtt>=1.6.1
pyyaml>=6.0
watchdog>=3.0.0
python-dsmr>=0.30  # For DSMR support
pyserial>=3.5
